refactro categoryCommands

+++ Restino.Application/Features/Categories/Commands/DeleteCategoryCommand/DeleteCategoryCommandHandler.cs
@@ -1,24 +1,26 @@
 ﻿using AutoMapper;
 using MediatR;
 using Restino.Application.Contracts.Persistance;
+using Restino.Application.Exceptions;
 using Restino.Domain.Entities;
 
 namespace Restino.Application.Features.Categories.Commands.DeleteCategoryCommand
 {
     public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand>
     {
-        private readonly IMapper _mapper;
         private readonly IAsyncRepository<Category> _categoryRepository;
 
-        public DeleteCategoryCommandHandler(IMapper mapper, IAsyncRepository<Category> categoryRepository)
+        public DeleteCategoryCommandHandler(IAsyncRepository<Category> categoryRepository)
         {
             _categoryRepository = categoryRepository;
-            _mapper = mapper;
         }
 
         public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
             var categoryToDelete = await _categoryRepository.GetByIdAsync(request.Id);
+
+            if (categoryToDelete == null)
+                throw new NotFoundException(nameof(Category), request.Id);
 
             await _categoryRepository.DeleteAsync(categoryToDelete);
